+++ e2e_playwright/st_chat_input_test.py
@@ -577,7 +577,7 @@ def test_uploads_and_deletes_single_file(
 
     file_upload_helper(themed_app, chat_input, [file1])
 
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
     expect(uploaded_files.get_by_text(file_name1)).to_be_visible()
     uploaded_files.scroll_into_view_if_needed()
 
@@ -593,12 +593,12 @@ def test_uploads_and_deletes_single_file(
     expect(uploaded_files.get_by_text(file_name2)).to_be_visible()
 
     # Delete the uploaded file
-    uploaded_files.get_by_test_id("stChatInputDeleteBtn").first.click()
+    uploaded_files.get_by_test_id("stFileChipDeleteBtn").first.click()
 
     wait_for_app_run(themed_app)
 
     # After deletion, the uploaded files container should not be visible
-    expect(chat_input.get_by_test_id("stChatUploadedFiles")).not_to_be_visible()
+    expect(chat_input.get_by_test_id("stFileChips")).not_to_be_visible()
 
 
 @use_chat_input("multiple_files")
@@ -621,7 +621,7 @@ def test_uploads_and_deletes_multiple_files(
 
     file_upload_helper(app, chat_input, files)
 
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
 
     # Wait for file names to be visible before taking snapshot
     expect(uploaded_files.get_by_text(file_name1)).to_be_visible()
@@ -632,15 +632,15 @@ def test_uploads_and_deletes_multiple_files(
 
     assert_snapshot(uploaded_files, name="st_chat_input-multiple_files_uploaded")
 
-    uploaded_file_names = uploaded_files.get_by_test_id("stChatInputFileName")
+    uploaded_file_names = uploaded_files.get_by_test_id("stFileChipName")
     expect(uploaded_file_names).to_have_count(2)
 
     # Delete one uploaded file
-    uploaded_files.get_by_test_id("stChatInputDeleteBtn").first.click()
+    uploaded_files.get_by_test_id("stFileChipDeleteBtn").first.click()
 
     wait_for_app_run(app)
 
-    uploaded_file_names = uploaded_files.get_by_test_id("stChatInputFileName")
+    uploaded_file_names = uploaded_files.get_by_test_id("stFileChipName")
     expect(uploaded_file_names).to_have_count(1)
 
     expect(uploaded_file_names).to_have_text(files[1]["name"], use_inner_text=True)
@@ -665,11 +665,11 @@ def test_file_upload_error_message_disallowed_files(
     )
 
     chat_input = get_element_by_key(themed_app, "single_file")
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
     expect(uploaded_files.get_by_text(file_name1)).to_be_visible()
 
     # Verify file chip has retry attributes (all errors are retryable)
-    file_chip = uploaded_files.get_by_test_id("stChatInputFile").first
+    file_chip = uploaded_files.get_by_test_id("stFileChip").first
     expect(file_chip).to_have_attribute("role", "button")
     expect(file_chip).to_have_attribute("tabindex", "0")
     expect(file_chip).to_have_attribute("title", "Click to retry upload")
@@ -707,13 +707,13 @@ def test_file_upload_error_message_file_too_large(app: Page):
     expect(chat_input).to_be_visible()
     file_upload_helper(app, chat_input, [file1])
 
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
     expect(uploaded_files).to_be_visible()
 
     # Verify the file appears in the uploaded files list
     expect(uploaded_files.get_by_text(file_name1)).to_be_visible()
 
-    uploaded_file = uploaded_files.get_by_test_id("stChatInputFile").first
+    uploaded_file = uploaded_files.get_by_test_id("stFileChip").first
     expect(uploaded_file).to_be_visible()
 
     uploaded_files.scroll_into_view_if_needed()
@@ -722,7 +722,7 @@ def test_file_upload_error_message_file_too_large(app: Page):
     reset_hovering(app)
 
     # Verify error message is displayed as a tooltip when hovering
-    file_chip = uploaded_files.get_by_test_id("stChatInputFile").first
+    file_chip = uploaded_files.get_by_test_id("stFileChip").first
     file_chip.hover()
     error_tooltip = app.get_by_test_id("stTooltipErrorContent").first
     expect(error_tooltip).to_be_visible()
@@ -1120,7 +1120,7 @@ def test_audio_input_visual_states(app: Page, assert_snapshot: ImageCompareFunct
     file_upload_helper(app, chat_input_with_files, [file])
 
     # Snapshot: Audio button + uploaded files
-    uploaded_files = chat_input_with_files.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input_with_files.get_by_test_id("stFileChips").first
     expect(uploaded_files).to_be_visible()
     assert_snapshot(
         chat_input_with_files, name="st_chat_input-audio_with_uploaded_files"
@@ -1226,6 +1226,43 @@ def test_audio_state_transitions_and_edge_cases(app: Page):
         expect(cancel_button).not_to_be_visible()
         expect(mic_button).to_be_visible()
         expect(mic_button).to_be_enabled()
+
+
+@use_chat_input("audio_with_files")
+@pytest.mark.skip_browser("webkit")  # Webkit CI audio permission issue
+def test_audio_with_all_features_combined(app: Page):
+    """Test audio with text and files all together (maximum complexity)."""
+    grant_microphone_permissions(app)
+
+    chat_input = get_element_by_key(app, "audio_with_files")
+    chat_input.scroll_into_view_if_needed()
+
+    # Upload files first
+    file1 = FilePayload(name="file1.txt", mimeType="text/plain", buffer=b"content1")
+    file2 = FilePayload(name="file2.txt", mimeType="text/plain", buffer=b"content2")
+    file_upload_helper(app, chat_input, [file1, file2])
+
+    # Verify files uploaded
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
+    expect(uploaded_files.get_by_text("file1.txt")).to_be_visible()
+    expect(uploaded_files.get_by_text("file2.txt")).to_be_visible()
+
+    # Add text after files are uploaded
+    textarea = chat_input.locator("textarea").first
+    textarea.fill("Message with everything")
+
+    # Record and submit audio (this submits everything together)
+    record_audio_in_chat_input(app, chat_input, duration_ms=1000)
+
+    # Verify text, audio, and files were all submitted successfully
+    expect_chat_input_value_contains_text(
+        app, "audio_with_files", "Message with everything"
+    )
+    expect_chat_input_value_contains_audio(app, "audio_with_files")
+    expect_chat_input_value_contains_files(app, "audio_with_files", 2)
+
+    # Verify textarea is cleared after submission
+    expect(textarea).to_have_value("")
 
 
 @use_chat_input("audio_column")
@@ -1420,12 +1457,12 @@ def upload_single_file_and_snapshot(
     """Helper to upload a single file and take a snapshot of just that file chip."""
     file_upload_helper(app, chat_input, [file])
 
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
-    file_chip = uploaded_files.get_by_test_id("stChatInputFile").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
+    file_chip = uploaded_files.get_by_test_id("stFileChip").first
     expect(file_chip).to_be_visible()
 
     # Verify title attribute contains full filename (for native tooltip on hover)
-    filename_element = uploaded_files.get_by_test_id("stChatInputFileName").first
+    filename_element = uploaded_files.get_by_test_id("stFileChipName").first
     expect(filename_element).to_have_attribute("title", file["name"])
 
     reset_hovering(app)
@@ -1433,7 +1470,7 @@ def upload_single_file_and_snapshot(
     assert_snapshot(file_chip, name=snapshot_name)
 
     # Delete the file to reset for next test
-    uploaded_files.get_by_test_id("stChatInputDeleteBtn").first.click()
+    uploaded_files.get_by_test_id("stFileChipDeleteBtn").first.click()
     wait_for_app_run(app, 500)
 
 
@@ -1533,8 +1570,8 @@ def test_file_upload_retry_click_success(app: Page):
         file_upload_helper(app, chat_input, [file])
 
         # Wait for error state to appear
-        uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
-        file_chip = uploaded_files.get_by_test_id("stChatInputFile").first
+        uploaded_files = chat_input.get_by_test_id("stFileChips").first
+        file_chip = uploaded_files.get_by_test_id("stFileChip").first
         expect(file_chip).to_be_visible()
 
         # Verify file is in error state with retry attributes
@@ -1555,7 +1592,7 @@ def test_file_upload_retry_click_success(app: Page):
         expect(tooltip_wrapper).not_to_be_visible(timeout=5000)
 
         # Verify file is now in uploaded state (shows size instead of error)
-        file_size = uploaded_files.get_by_test_id("stChatInputFileName").first
+        file_size = uploaded_files.get_by_test_id("stFileChipName").first
         expect(file_size).to_be_visible()
 
     finally:
@@ -1593,16 +1630,16 @@ def test_upload_button_works_after_upload_and_delete(app: Page):
     wait_for_app_run(app, 500)
 
     # Verify the file was uploaded
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
     expect(uploaded_files).to_be_visible()
     expect(uploaded_files.get_by_text(first_file_name)).to_be_visible()
 
     # Step 2: Delete the uploaded file
-    uploaded_files.get_by_test_id("stChatInputDeleteBtn").first.click()
+    uploaded_files.get_by_test_id("stFileChipDeleteBtn").first.click()
     wait_for_app_run(app, 500)
 
     # Verify the file was deleted
-    expect(chat_input.get_by_test_id("stChatUploadedFiles")).not_to_be_visible()
+    expect(chat_input.get_by_test_id("stFileChips")).not_to_be_visible()
 
     # Step 3: Verify the upload button still works after the upload + delete cycle
     # This is the key assertion - without the fix, the button would be unresponsive
@@ -1623,7 +1660,7 @@ def test_upload_button_works_after_upload_and_delete(app: Page):
     wait_for_app_run(app, 500)
 
     # Verify the new file was uploaded successfully
-    uploaded_files = chat_input.get_by_test_id("stChatUploadedFiles").first
+    uploaded_files = chat_input.get_by_test_id("stFileChips").first
     expect(uploaded_files).to_be_visible()
     expect(uploaded_files.get_by_text(second_file_name)).to_be_visible()
