+++ frontend/lib/src/components/widgets/ChatInput/fileUpload/styled-components.ts
@@ -15,8 +15,6 @@
  */
 import styled from "@emotion/styled"
 
-import { FileStatus } from "~lib/components/shared/UploadedFile/UploadFileInfo"
-
 // A transparent dropzone overlay that covers the ContentArea
 export const StyledChatFileUploadDropzone = styled.div(({ theme }) => ({
   backgroundColor: theme.colors.transparent,
@@ -51,161 +49,3 @@ export const StyledFileUploadButton = styled.div<StyledFileUploadButtonProps>(
     pointerEvents: disabled ? "none" : "auto",
   })
 )
-
-export const StyledChatUploadedFiles = styled.div(({ theme }) => ({
-  lineHeight: theme.lineHeights.tight,
-}))
-
-export const StyledUploadedChatFileList = styled.div(({ theme }) => ({
-  display: "flex",
-  flexWrap: "wrap",
-  gap: theme.spacing.sm, // Figma: 8px gap between file chips
-}))
-
-export const StyledUploadedChatFileListItem = styled.div({
-  flex: "0 0 auto",
-  maxWidth: "100%",
-})
-
-export interface StyledChatUploadedFileProps {
-  isError?: boolean
-  isClickable?: boolean
-}
-
-export const StyledChatUploadedFile = styled.div<StyledChatUploadedFileProps>(
-  ({ theme, isError, isClickable }) => ({
-    position: "relative",
-    display: "inline-flex",
-    alignItems: "center",
-    width: "fit-content",
-    minWidth: "9rem",
-    maxWidth: "100%",
-    backgroundColor: isError
-      ? theme.colors.redBackgroundColor
-      : theme.colors.bgColor,
-    padding: theme.spacing.twoXS,
-    paddingRight: theme.spacing.twoXL, // Extra padding for absolute positioned X button
-    borderRadius: theme.radii.default,
-    gap: theme.spacing.sm,
-    cursor: isClickable ? "pointer" : "default",
-  })
-)
-
-// Container for filename and size stacked vertically
-export const StyledChatUploadedFileInfo = styled.div({
-  display: "flex",
-  flexDirection: "column",
-  minWidth: 0, // Allow text truncation
-})
-
-export interface StyledChatUploadedFileIconContainerProps {
-  fileStatus: "uploading" | "uploaded" | "error"
-}
-
-export const StyledChatUploadedFileIconContainer =
-  styled.div<StyledChatUploadedFileIconContainerProps>(
-    ({ theme, fileStatus }) => ({
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-      borderRadius: theme.radii.default,
-      width: theme.sizes.chatInputFileIconSize,
-      height: theme.sizes.chatInputFileIconSize,
-      flexShrink: 0,
-      overflow: "hidden",
-      ...(fileStatus === "uploaded" && {
-        backgroundColor: theme.colors.bodyText,
-        color: theme.colors.bgColor,
-      }),
-      ...(fileStatus === "uploading" && {
-        backgroundColor: theme.colors.fadedText10,
-        color: theme.colors.fadedText60,
-      }),
-      ...(fileStatus === "error" && {
-        backgroundColor: theme.colors.redBackgroundColor,
-        color: theme.colors.redTextColor,
-      }),
-    })
-  )
-
-export const StyledChatUploadedFileImagePreview = styled.img({
-  width: "100%",
-  height: "100%",
-  objectFit: "cover",
-})
-
-export interface StyledChatUploadedFileStatusProps {
-  fileStatus: FileStatus
-}
-
-export const StyledChatUploadedFileName =
-  styled.div<StyledChatUploadedFileStatusProps>(({ theme, fileStatus }) => ({
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-    whiteSpace: "nowrap",
-    color:
-      fileStatus.type === "uploading"
-        ? theme.colors.fadedText60
-        : theme.colors.bodyText,
-  }))
-
-export const StyledChatUploadedFileSize = styled.div(({ theme }) => ({
-  color: theme.colors.fadedText60,
-  fontSize: theme.fontSizes.sm,
-}))
-
-export interface StyledChatUploadedFileDeleteButtonProps {
-  isError?: boolean
-}
-
-export const StyledChatUploadedFileDeleteButton =
-  styled.small<StyledChatUploadedFileDeleteButtonProps>(
-    ({ theme, isError }) => ({
-      position: "absolute",
-      top: theme.spacing.twoXS,
-      right: theme.spacing.twoXS,
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-      lineHeight: 0,
-      // Circular background for the X button
-      "& button": {
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        width: "fit-content",
-        height: "fit-content",
-        minHeight: "unset",
-        minWidth: "unset",
-        maxHeight: "unset",
-        maxWidth: "unset",
-        borderRadius: "50%",
-        backgroundColor: "transparent",
-        color: isError ? theme.colors.redTextColor : theme.colors.fadedText60,
-        padding: 0,
-        overflow: "hidden",
-        boxSizing: "border-box",
-        lineHeight: 0,
-        "&:hover": {
-          backgroundColor: "transparent",
-          color: isError ? theme.colors.redColor : theme.colors.bodyText,
-        },
-      },
-    })
-  )
-
-/* eslint-disable streamlit-custom/no-hardcoded-theme-values */
-// Visually hidden but accessible to screen readers
-// Uses standard CSS visually-hidden pattern (hardcoded values required)
-export const StyledVisuallyHidden = styled.span({
-  position: "absolute",
-  width: "1px",
-  height: "1px",
-  padding: 0,
-  margin: "-1px",
-  overflow: "hidden",
-  clip: "rect(0, 0, 0, 0)",
-  whiteSpace: "nowrap",
-  border: 0,
-})
-/* eslint-enable streamlit-custom/no-hardcoded-theme-values */
